instance and nucleus tests in correct place
