+++ core/src/instance.rs
@@ -35,10 +35,12 @@ impl Observer {
 pub static DISPATCH_WITHOUT_CHANNELS: &str = "dispatch called without channels open";
 
 impl Instance {
+    /// get a clone of the action channel
     pub fn action_channel(&self) -> Sender<ActionWrapper> {
         self.action_channel.clone()
     }
 
+    /// get a clone of the observer channel
     pub fn observer_channel(&self) -> Sender<Observer> {
         self.observer_channel.clone()
     }
@@ -209,4 +211,152 @@ pub fn dispatch_action(
         .send(wrapper.clone())
         .unwrap_or_else(|_| panic!(DISPATCH_WITHOUT_CHANNELS));
     wrapper
+}
+
+#[cfg(test)]
+pub mod tests {
+    extern crate test_utils;
+    use std::thread::sleep;
+    use nucleus::Action::InitApplication;
+    use state::Action::Nucleus;
+    use std::sync::mpsc::channel;
+    use state::State;
+    use super::Instance;
+    use holochain_dna::Dna;
+    use std::time::Duration;
+    use holochain_dna::zome::capabilities::ReservedCapabilityNames;
+
+    /// create a test instance
+    pub fn test_instance(dna: Dna) -> Instance {
+        // Create instance and plug in our DNA
+        let mut instance = Instance::new();
+        let action = Nucleus(InitApplication(dna.clone()));
+        instance.start_action_loop();
+        instance.dispatch_and_wait(action.clone());
+        assert_eq!(instance.state().nucleus().dna(), Some(dna));
+
+        // Wait for Init to finish
+        while instance.state().history.len() < 4 {
+            // TODO - #21
+            // This println! should be converted to either a call to the app logger, or to the core debug log.
+            println!("Waiting... {}", instance.state().history.len());
+            sleep(Duration::from_millis(10))
+        }
+
+        instance
+    }
+
+    /// This test shows how to call dispatch with a closure that should run
+    /// when the action results in a state change.  Note that the observer closure
+    /// needs to return a boolean to indicate that it has successfully observed what
+    /// it intends to observe.  It will keep getting called as the state changes until
+    /// it returns true.
+    /// Note also that for this test we create a channel to send something (in this case
+    /// the dna) back over, just so that the test will block until the closure is successfully
+    /// run and the assert will actually run.  If we put the assert inside the closure
+    /// the test thread could complete before the closure was called.
+
+    #[test]
+    fn can_dispatch_with_observer() {
+        let mut instance = Instance::new();
+        instance.start_action_loop();
+
+        let dna = Dna::new();
+        let (sender, receiver) = channel();
+        instance.dispatch_with_observer(
+            Nucleus(InitApplication(dna.clone())),
+            move |state: &State| match state.nucleus().dna() {
+                Some(dna) => {
+                    sender.send(dna).expect("test channel must be open");
+                    return true;
+                }
+                None => return false,
+            },
+        );
+
+        let stored_dna = receiver.recv().unwrap();
+
+        assert_eq!(dna, stored_dna);
+    }
+
+    #[test]
+    fn can_dispatch_and_wait() {
+        let mut instance = Instance::new();
+        assert_eq!(instance.state().nucleus().dna(), None);
+        assert_eq!(
+            instance.state().nucleus().status(),
+            ::nucleus::NucleusStatus::New
+        );
+
+        let dna = Dna::new();
+        let action = Nucleus(InitApplication(dna.clone()));
+        instance.start_action_loop();
+
+        // the initial state is not intialized
+        assert!(instance.state().nucleus().has_initialized() == false);
+
+        instance.dispatch_and_wait(action.clone());
+        assert_eq!(instance.state().nucleus().dna(), Some(dna));
+
+        // Wait for Init to finish
+        while instance.state().history.len() < 2 {
+            println!("Waiting... {}", instance.state().history.len());
+            sleep(Duration::from_millis(10));
+        }
+        assert!(instance.state().nucleus().has_initialized());
+    }
+
+    #[test]
+    fn test_genesis_ok() {
+        let dna = test_utils::create_test_dna_with_wat(
+            "test_zome".to_string(),
+            ReservedCapabilityNames::LifeCycle.as_str().to_string(),
+            Some(
+                r#"
+            (module
+                (memory (;0;) 17)
+                (func (export "genesis_dispatch") (param $p0 i32) (param $p1 i32) (result i32)
+                    i32.const 0
+                )
+                (data (i32.const 0)
+                    ""
+                )
+                (export "memory" (memory 0))
+            )
+        "#,
+            ),
+        );
+
+        let instance = test_instance(dna);
+
+        assert_eq!(instance.state().history.len(), 4);
+        assert!(instance.state().nucleus().has_initialized());
+    }
+
+    #[test]
+    fn test_genesis_err() {
+        let dna = test_utils::create_test_dna_with_wat(
+            "test_zome".to_string(),
+            ReservedCapabilityNames::LifeCycle.as_str().to_string(),
+            Some(
+                r#"
+            (module
+                (memory (;0;) 17)
+                (func (export "genesis_dispatch") (param $p0 i32) (param $p1 i32) (result i32)
+                    i32.const 4
+                )
+                (data (i32.const 0)
+                    "1337"
+                )
+                (export "memory" (memory 0))
+            )
+        "#,
+            ),
+        );
+
+        let instance = test_instance(dna);
+
+        assert_eq!(instance.state().history.len(), 4);
+        assert!(instance.state().nucleus().has_initialized() == false);
+    }
 }
