+++ core/src/nucleus/mod.rs
@@ -465,11 +465,14 @@ pub fn reduce(
 
 #[cfg(test)]
 mod tests {
+    extern crate test_utils;
+    use ::instance::tests::test_instance;
     use super::{
         super::{nucleus::Action::*, state::Action::*},
         *,
     };
     use std::sync::mpsc::channel;
+    use ::instance::Instance;
 
     #[test]
     fn can_instantiate_nucleus_state() {
@@ -582,5 +585,109 @@ mod tests {
         let (tx_observer, _observer) = channel::<Observer>();
         let reduced_nucleus = reduce(nucleus.clone(), &action, &sender, &tx_observer);
         assert_eq!(nucleus, reduced_nucleus);
+    }
+
+    #[test]
+    fn call_ribosome_function() {
+        let dna = test_utils::create_test_dna_with_wat(
+            "test_zome".to_string(),
+            "test_cap".to_string(),
+            None,
+        );
+        let mut instance = test_instance(dna);
+
+        // Create zome function call
+        let call = FunctionCall::new("test_zome", "test_cap", "main", "");
+
+        let result = call_and_wait_for_result(call, &mut instance);
+        match result {
+            // Result 1337 from WASM (as string)
+            Ok(val) => assert_eq!(val, "1337"),
+            Err(err) => assert_eq!(err, HolochainError::InstanceActive),
+            //Err(_) => assert!(false),
+        }
+    }
+
+    #[test]
+    fn call_ribosome_wrong_dna() {
+        let mut instance = Instance::new();
+        instance.start_action_loop();
+
+        let call = FunctionCall::new("test_zome", "test_cap", "main", "{}");
+        let result = call_and_wait_for_result(call, &mut instance);
+
+        match result {
+            Err(HolochainError::DnaMissing) => {}
+            _ => assert!(false),
+        }
+    }
+
+    #[test]
+    fn call_ribosome_wrong_function() {
+        let dna = test_utils::create_test_dna_with_wat(
+            "test_zome".to_string(),
+            "test_cap".to_string(),
+            None,
+        );
+        let mut instance = test_instance(dna);
+
+        // Create zome function call:
+        let call = FunctionCall::new("test_zome", "test_cap", "xxx", "{}");
+
+        let result = call_and_wait_for_result(call, &mut instance);
+
+        match result {
+            Err(HolochainError::ErrorGeneric(err)) => {
+                assert_eq!(err, "Function: Module doesn\'t have export xxx_dispatch")
+            }
+            _ => assert!(false),
+        }
+    }
+
+    #[test]
+    fn call_wrong_ribosome_function() {
+        let dna = test_utils::create_test_dna_with_wat(
+            "test_zome".to_string(),
+            "test_cap".to_string(),
+            None,
+        );
+        let mut instance = test_instance(dna);
+
+        // Create bad zome function call
+        let call = FunctionCall::new("xxx", "test_cap", "main", "{}");
+
+        let result = call_and_wait_for_result(call, &mut instance);
+
+        match result {
+            Err(HolochainError::ZomeNotFound(err)) => assert_eq!(err, "Zome 'xxx' not found"),
+            _ => assert!(false),
+        }
+
+        // Create bad capability function call
+        let call = FunctionCall::new("test_zome", "xxx", "main", "{}");
+
+        let result = call_and_wait_for_result(call, &mut instance);
+
+        match result {
+            Err(HolochainError::CapabilityNotFound(err)) => {
+                assert_eq!(err, "Capability 'xxx' not found in Zome 'test_zome'")
+            }
+            _ => assert!(false),
+        }
+    }
+
+    #[test]
+    fn test_missing_genesis() {
+        let mut dna = test_utils::create_test_dna_with_wat(
+            "test_zome".to_string(),
+            "test_cap".to_string(),
+            None,
+        );
+        dna.zomes[0].capabilities[0].name = ReservedCapabilityNames::LifeCycle.as_str().to_string();
+
+        let instance = test_instance(dna);
+
+        assert_eq!(instance.state().history.len(), 4);
+        assert!(instance.state().nucleus().has_initialized());
     }
 }
